test: split store reconciliation tests
